@@ -6,7 +6,7 @@
   "scripts": {
     "build": "tsc -p .",
     "start": "npm run build && node dist",
-    "dev": "tsc -w" ,
+    "dev": "tsc -w",
     "devStart": "nodemon dist"
   },
   "author": {
@@ -24,7 +24,11 @@
   },
   "devDependencies": {
     "@types/node": "^16.11.12",
+    "@typescript-eslint/eslint-plugin": "^5.27.0",
+    "@typescript-eslint/parser": "^5.27.0",
+    "eslint": "^8.16.0",
+    "eslint-config-airbnb-typescript": "^17.0.0",
     "nodemon": "^2.0.15",
-    "typescript": "^4.5.3"
+    "typescript": "^4.7.2"
   }
 }
